+++ Assets/ProceduralStages/Terrain/Themes/Theme.cs
@@ -10,6 +10,7 @@ namespace ProceduralStages
     {
         Random,
         LegacyRandom,
-        Desert
+        Desert,
+        Snow
     }
 }