+++ Assets/ProceduralStages/Terrain/MapGenerator.cs
@@ -158,7 +158,7 @@ namespace ProceduralStages
                     ? selection.Evaluate(rng.nextNormalizedFloat)
                     : TerrainType.OpenCaves;
 
-                terrainType = TerrainType.Moon;
+                //terrainType = TerrainType.Moon;
 
             }